requirements.txt
aioconsole==0.7.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
eval-type-backport==0.1.3
exceptiongroup==1.2.1
frozenlist==1.4.1
grpcio==1.63.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-path==0.3.2
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
MarkupSafe==2.1.5
more-itertools==10.2.0
motor==3.4.0
multidict==6.0.5
nest-asyncio==1.6.0
openai==1.24.1
openapi-core==0.19.1
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
packaging==24.0
parse==1.20.1
pathable==0.4.3
prance==23.6.21.0
pybars4==0.9.13
pydantic==2.7.1
pydantic_core==2.18.2
PyMeta3==0.5.1
pymongo==4.7.1
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.31.1
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scipy==1.13.0
semantic-kernel==0.9.6b1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
yarl==1.9.4
